Safari fullscreen video text track cues illegible size https://bugs.webkit.org/show_bug.cgi?id=290576 rdar://145149369
Reviewed by NOBODY (OOPS!).

On AVKit fullscreen on iPhone, when VTT captions are put into fullscreen, they shrink down substantially.
This is because we multiply the videoBox's width and height by the device scale factor to represent the
increased number of pixels in the fullscreen resolution, however this messes up the later shrinking calculation
that we do on the captions.

* Source/WebCore/html/shadow/MediaControlTextTrackContainerElement.cpp:
(WebCore::MediaControlTextTrackContainerElement::updateVideoDisplaySize):