Quirk premierleague.com to fix fullscreen video position after return from PIP on iPad https://bugs.webkit.org/show_bug.cgi?id=293436 rdar://138050196
Reviewed by NOBODY (OOPS!).

PremierLeague.com has a <div> around the <video> with an attribute called data-fullscreen.
This attribute controls the div's styling. When the user exits pip and the video transitions
back to fullscreen, the site does not set data-fullscreen to true, so their inline styling
is applied instead, and the video moves to the top of the viewport.

While waiting for the site to fix their bug, we can apply a quirk that corrects the css style
on the div.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::shouldInjectCSSInFullscreen const):
(WebCore::handlePremierLeagueQuirks):
* Source/WebCore/page/Quirks.h:
* Source/WebCore/page/QuirksData.h:
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::adjustForSiteSpecificQuirks const):